requirements.txt
celery==5.2.6
Django==4.0.2
djangorestframework==3.13.1
numpy==1.22.1
pandas==1.4.0
python-binance==1.0.15
ta==0.9.0
websocket-client==1.2.3
psycopg2
pandas-ta
gunicorn
redis
